+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Word converter applying other word converters in series. The output for the first word converter is used as input for the second word converter, etc.
 
 ### Changed
-- Nothing
+- Word converters have been renamed word formatters.
+- The convert method of word formatters (former word converters) has been renamed apply.
 
 ### Fixed
 - Nothing